msm8996-common: Do not enable updatable APEXes on these targets
This reverts commit 106ac21a9281f33472a8481d77d5256756ac83ef.

It can't really be enabled for z2_plus and z2_row, as apexd requires /data/apex
to be available right after boot to update all the Android modules.
As in all FDE devices, /data/apex is encrypted until the device
is unlocked by the user, rendering APEX basically useless as only
the system APEX variants will be loaded at boot.

Change-Id: I0649862e5c5b2c1a1cb597914a11df79b3e01957